+++ sqlite3/lib/open.dart
@@ -1,5 +1,7 @@
 /// Utils to open a [DynamicLibrary] on platforms that aren't supported by
 /// default.
+///
+/// {@category native}
 library;
 
 import 'dart:ffi';