+++ sqlite3/lib/sqlite3.dart
@@ -1,4 +1,6 @@
 /// Dart bindings to `sqlite3`.
+///
+/// {@category native}
 library;
 
 // Hide common base classes that have more specific ffi-APIs.